feat: implemented copy clipboard event added password crate added lib.rs file
