+++ src/clipboard.rs
@@ -4,12 +4,12 @@ pub struct Clipboard{
     password : String,
 }
 
-pub impl Clipboard{
-    fn new(_password: String) -> Self{
+impl Clipboard{
+    pub fn new(_password: String) -> Self{
         Self{password:_password}
     }
     
-    pub fn Copy(&self) {
+    pub fn copy(&self) {
         let mut ctx: ClipboardContext = ClipboardProvider::new().unwrap();
         ctx.set_contents(self.password.to_owned()).unwrap();
     }
